admin: extract kafka_connections_service
To encapsulate the logic of konnection gathering.

The code is only moved, the only changes are extracting a
`get_effective_limit` function for later reuse, and some minor fixups to
make the code compile (e.g.; type name adjustments due to moving across
namespaces).
