+++ src/v/redpanda/admin/services/BUILD
@@ -9,14 +9,10 @@ redpanda_cc_library(
     deps = [
         "//proto/redpanda/core/admin/v2:broker_redpanda_proto",
         "//proto/redpanda/core/admin/v2:kafka_connections_redpanda_proto",
-        "//src/v/container:priority_queue",
         "//src/v/features",
-        "//src/v/kafka/server",
-        "//src/v/redpanda/admin:aip_filter",
-        "//src/v/redpanda/admin:aip_ordering",
+        "//src/v/redpanda/admin:kafka_connections_service",
         "//src/v/redpanda/admin/proxy:client",
         "//src/v/serde/protobuf:rpc",
-        "//src/v/ssx:async_algorithm",
         "//src/v/version",
         "@seastar",
     ],
